+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/ConjunctionAssetFilter.java
@@ -9,7 +9,6 @@ import org.immutables.value.Value;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @WealthdraftImmutableStyle