+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/TagAssetFilter.java
@@ -9,9 +9,7 @@ import com.strangegrotto.wealthdraft.assets.definition.CustomTagDefinition;
 import com.strangegrotto.wealthdraft.assets.definition.IntrinsicAssetTag;
 import org.immutables.value.Value;
 
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @WealthdraftImmutableStyle
@@ -27,7 +25,7 @@ public abstract class TagAssetFilter implements AssetFilter {
     public abstract String getValue();
 
     @Override
-    public Map<String, Asset> apply(Map<String, Asset> input) {
+    public Map<String, Asset> apply(Map<String, AssetFilter> allFilters, Map<String, Asset> input) {
         var tagName = getTag();
         var tagValue = getValue();
         return input.entrySet().stream()
@@ -37,6 +35,11 @@ public abstract class TagAssetFilter implements AssetFilter {
                     return assetTags.containsKey(tagName) && tagValue.equals(assetTags.get(tagName));
                 })
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    }
+
+    @Override
+    public Optional<List<String>> checkForCycles(Map<String, AssetFilter> allFilters, LinkedHashSet<String> parentFilters) {
+        return Optional.empty();
     }
 
     @Value.Check