+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/ConjunctionAssetFilterTest.java
@@ -9,7 +9,6 @@ import org.junit.Test;
 
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 public class ConjunctionAssetFilterTest {
     @Test
@@ -55,7 +54,7 @@ public class ConjunctionAssetFilterTest {
                 needle2Filter
         ));
 
-        var result = conjunctiveFilter.apply(haystack);
+        var result = conjunctiveFilter.apply(Map.of(), haystack);
 
         var expected = Map.of(
                 matchingAssetId, haystack.get(matchingAssetId)