+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/TagAssetFilterTest.java
@@ -7,7 +7,6 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.Map;
-import java.util.Set;
 
 public class TagAssetFilterTest {
     @Test
@@ -37,7 +36,7 @@ public class TagAssetFilterTest {
         );
 
         var filter = ImmTagAssetFilter.of(customTags, selectedTagName, selectedTagValue);
-        var result = filter.apply(assets);
+        var result = filter.apply(Map.of(), assets);
         var expected = Map.of(
                 matchingAssetId1, assets.get(matchingAssetId1),
                 matchingAssetId2, assets.get(matchingAssetId2)